Jebena state mašina radi, govno posrano.

+++ Tangram/Vec2.h
@@ -13,7 +13,7 @@ class Vec2 {
         void setX(int x) { m_x = x; }
         void setY(int y) { m_y = y; }
 
-        int length() { return sqrt(m_x * m_x + m_y * m_y); }
+        int length() { return (int)sqrt(m_x * m_x + m_y * m_y); }
 
         Vec2 operator+(const Vec2& v2) const{return Vec2(m_x + v2.m_x, m_y + v2.m_y);}
         friend Vec2& operator+=(Vec2& v1, const Vec2& v2)
